requirements.txt
msal
requests
python-dotenv
google-genai
tiktoken
azure-communication-email
markdown
openai